Renamed nftAsset >> NFTAsset
