+++ models/NFTAsset.ts
