+++ src/fragments/DropDown.Collections.MintForm.tsx
@@ -3,7 +3,7 @@ import classes from "../components/styles/RequestForm.module.scss";
 
 import React, { useEffect } from "react";
 
-import { NFTAsset } from "../../models/nftAsset";
+import { NFTAsset } from "../../models/NFTAsset";
 import { Select, MenuItem, InputLabel, FormControl } from "@mui/material";
 
 type Props =
